mobile.css
This file is the navigation bar css framework for several different viewing browser sizes, so basically just for different media devices like an iPad/Tablet, a phone, and a normal pc. 
The nav bar will change from the three, there may be a few corrections yet needed to be made but this is apart of my lab assignment for class due December 6th, leeeeggggoo!